# Docusaurus v3.6 - enable Docusaurus Faster