--- a/demo/package.json
+++ b/demo/package.json
@@ -21,10 +21,10 @@
     "re-gen": "yarn clean-all && yarn gen-all"
   },
   "dependencies": {
-    "@docusaurus/core": "3.5.2-canary-6122",
-    "@docusaurus/faster": "3.5.2-canary-6122",
-    "@docusaurus/plugin-google-gtag": "3.5.2-canary-6122",
-    "@docusaurus/preset-classic": "3.5.2-canary-6122",
+    "@docusaurus/core": "3.6.0",
+    "@docusaurus/faster": "3.6.0",
+    "@docusaurus/plugin-google-gtag": "3.6.0",
+    "@docusaurus/preset-classic": "3.6.0",
     "clsx": "^1.1.1",
     "docusaurus-plugin-openapi-docs": "^4.1.0",
     "docusaurus-theme-openapi-docs": "^4.1.0",