--- a/packages/docusaurus-plugin-openapi-docs/package.json
+++ b/packages/docusaurus-plugin-openapi-docs/package.json
@@ -28,7 +28,7 @@
     "watch": "tsc --watch"
   },
   "devDependencies": {
-    "@docusaurus/types": "^3.5.0",
+    "@docusaurus/types": "3.5.2-canary-6122",
     "@types/fs-extra": "^9.0.13",
     "@types/json-pointer": "^1.0.31",
     "@types/json-schema": "^7.0.9",
@@ -38,9 +38,9 @@
   },
   "dependencies": {
     "@apidevtools/json-schema-ref-parser": "^11.5.4",
-    "@docusaurus/plugin-content-docs": "^3.5.0",
-    "@docusaurus/utils": "^3.5.0",
-    "@docusaurus/utils-validation": "^3.5.0",
+    "@docusaurus/plugin-content-docs": "3.5.2-canary-6122",
+    "@docusaurus/utils": "3.5.2-canary-6122",
+    "@docusaurus/utils-validation": "3.5.2-canary-6122",
     "@redocly/openapi-core": "^1.10.5",
     "allof-merge": "^0.6.6",
     "chalk": "^4.1.2",