@slorber 2024-11-04
v3.6 is out: https://docusaurus.io/blog/releases/3.6

PR ready to review

Note the HTML minifier reports a few errors on the demo site, you might want to investigate those so that plugin users do not get invalid markup + the console errors.